Adjust timeouts and fix code coverage
